последние изменения также как предыдущеее

+++ Bank_End/Models/Account.cs
@@ -8,7 +8,7 @@ namespace Bank_End.Models
 {
     public class Account
     {
-        public int AccountID { get; set; }
+        public long AccountID { get; set; }
         public decimal Balanse { get; set; }
         public int Currency { get; set; }
         public User Owner { get; set; }
