+++ Bank_End/Models/Credit.cs
@@ -8,7 +8,7 @@ namespace Bank_End.Models
 {
     public class Credit
     {
-        public int CreditID { get; set; }
+        public long CreditID { get; set; }
         public int Regularity { get; set; }
         public decimal CreditRate { get; set; }
         public DateTime IssuanceDate { get; set; }
