+++ Bank_End/Models/Deposit.cs
@@ -8,7 +8,7 @@ namespace Bank_End.Models
 {
     public class Deposit
     {   // Я знаю, что правильно deposit*, позндно заметил ошибку, много где менять ещё надо  
-        public int DepositID { get; set; }
+        public long DepositID { get; set; }
         public int Regularity { get; set; }
         public decimal DepositRate { get; set; }
         public DateTime IssuanceDate { get; set; }
